Rename the pair-only sierra-miro merger rule to reflect what it does
