+++ pipeline/merger/src/main/scala/uk/ac/wellcome/platform/merger/services/Merger.scala
@@ -4,7 +4,7 @@ import uk.ac.wellcome.models.work.internal._
 import uk.ac.wellcome.platform.merger.rules.MergerRule
 import uk.ac.wellcome.platform.merger.rules.physicaldigital.SierraPhysicalDigitalMergeRule
 import uk.ac.wellcome.platform.merger.rules.sierramets.SierraMetsMergerRule
-import uk.ac.wellcome.platform.merger.rules.singlepagemiro.SinglePageMiroMergeRule
+import uk.ac.wellcome.platform.merger.rules.sierramiropair.SierraMiroPairMergeRule
 
 class Merger(rules: List[MergerRule]) {
   def merge(works: Seq[TransformedBaseWork]): Seq[BaseWork] = {
@@ -17,5 +17,5 @@ object PlatformMerger
     extends Merger(
       List(
         SierraPhysicalDigitalMergeRule,
-        SinglePageMiroMergeRule,
+        SierraMiroPairMergeRule,
         SierraMetsMergerRule))
